NICE HASH STRATUM FIX-WITH DIFFICULTY SETTINGS

+++ proxy/config.go
@@ -36,7 +36,7 @@ type Proxy struct {
 	BehindReverseProxy   bool   `json:"behindReverseProxy"`
 	BlockRefreshInterval string `json:"blockRefreshInterval"`
 	Difficulty           int64  `json:"difficulty"`
-	DifficultyNiceHash   float64 `json:"difficultyNiceHash"`
+	DifficultyNiceHash   int64 `json:"difficultyNiceHash"`
 	StateUpdateInterval  string `json:"stateUpdateInterval"`
 	HashrateExpiration   string `json:"hashrateExpiration"`
